wip: setup.py
