+++ must3r/demo/gradio.py
@@ -11,7 +11,9 @@ import numpy as np
 import functools
 import trimesh
 import datetime
+import tempfile
 from scipy.spatial.transform import Rotation
+import matplotlib.pyplot as pl
 
 from must3r.demo.viser import ViserWrapper
 from must3r.demo.inference import *
@@ -23,13 +25,18 @@ from dust3r.utils.device import to_numpy
 from dust3r.utils.geometry import geotrf
 from dust3r.viz import add_scene_cam, CAM_COLORS, OPENGL, pts3d_to_trimesh, cat_meshes
 
+from must3r.model import *
 from must3r.model.blocks.layers import MEMORY_MODES
+from must3r.model.blocks.attention import has_xformers, toggle_memory_efficient_attention
 
 try:
     from pillow_heif import register_heif_opener  # noqa
     register_heif_opener()
 except ImportError:
     pass
+
+
+pl.ion()
 
 
 def get_args_parser():
@@ -458,3 +465,33 @@ def main_demo(tmpdirname, model, retrieval, device, image_size, server_name, ser
                                            local_pointmaps, cam_size, camera_conf_thr],
                                    outputs=outmodel)
     demo.launch(share=False, server_name=server_name, server_port=server_port)
+
+
+def main():
+    torch.backends.cuda.matmul.allow_tf32 = True  # for gpu >= Ampere and pytorch >= 1.12
+    parser = get_args_parser()
+    args = parser.parse_args()
+
+    toggle_memory_efficient_attention(enabled=has_xformers)
+
+    if args.tmp_dir is not None:
+        tmp_path = args.tmp_dir
+        os.makedirs(tmp_path, exist_ok=True)
+        tempfile.tempdir = tmp_path
+
+    if args.server_name is not None:
+        server_name = args.server_name
+    else:
+        server_name = '0.0.0.0' if args.local_network else '127.0.0.1'
+
+    weights_path = args.weights
+    model = load_model(weights_path, encoder=args.encoder, decoder=args.decoder, device=args.device,
+                       img_size=args.image_size, memory_mode=args.memory_mode, verbose=args.verbose)
+
+    # must3r will write the 3D model inside tmpdirname
+    with tempfile.TemporaryDirectory(suffix='dust3r_gradio_demo') as tmpdirname:
+        if args.verbose:
+            print('Outputing stuff in', tmpdirname)
+        main_demo(tmpdirname, model, args.retrieval, args.device, args.image_size,
+                  server_name, args.server_port, verbose=args.verbose, amp=args.amp, with_viser=args.viser,
+                  allow_local_files=args.allow_local_files)
